requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
transformers>=4.28.0
datasets>=2.11.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.15.0
hydra-core>=1.3.0
omegaconf>=2.3.0
tensorboard>=2.13.0
ipykernel>=6.0.0
jupyter>=1.0.0
tqdm>=4.65.0
pyarrow>=12.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pytest>=7.3.0
pre-commit>=3.3.0
kaggle>=1.5.0
opencv-python>=4.7.0
albumentations>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.1.0
optuna>=3.0.0
shap>=0.41.0
plotly>=5.13.0 